requirements.txt
py-clob-client>=0.17.0
python-dotenv>=1.0.0
twilio>=8.10.0
websockets>=12.0
aiohttp>=3.9.0
web3>=6.11.0
eth-account>=0.10.0
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
